mcp_agent/pyproject.toml
[project]
name = "mcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.12.4",
    "playwright>=1.55.0",
    "pydantic-ai[mcp]>=1.0.17",
    "python-dotenv>=1.1.1",
]
